feat: add notification badge and read status sync
